+++ src/app/routes/private/routes/integrantes.private.js
@@ -6,5 +6,6 @@ const integrantes = require('../../../models/integrantes')
 route.post('/criar', IntegrantesController.create)
 route.get('/listar-todos', IntegrantesController.getWithParams)
 route.get('/listar-um/:nome', IntegrantesController.getByName);
+route.put('/atualizar/:nome',IntegrantesController.update);
 
 module.exports = route