+++ CHANGELOG.md
@@ -36,3 +36,4 @@ Log de alterações da aplicação
 
 ### Versão 1.3
 * **Integrantes:** Alterando a model de integrante para aceitar email.
+* **Integrantes:** Adição da rota e metodo para editar integrante.